Changed subscribe_candlestick() interface

+++ crypto-ws-client/src/clients/bybit/bybit_linear_swap.rs
@@ -32,7 +32,7 @@ impl<'a> BBO for BybitLinearSwapWSClient<'a> {
     }
 }
 
-fn to_candlestick_raw_channel(pair: &str, interval: u32) -> String {
+fn to_candlestick_raw_channel(pair: &str, interval: usize) -> String {
     let interval_str = match interval {
         60 => "1",
         180 => "3",
